backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
supabase>=2.0.0
asyncpg>=0.29
python-dotenv>=1.0
httpx>=0.27
PyJWT[crypto]>=2.8
